requirements.txt
google-auth==2.28.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.120.0
psycopg2==2.9.9
